CFA : Ajout du panel Trust Ajout de confiance dans Upgrade Manager Modification de LossTrustRate en TrustRate afin de mieux comprendre son utilité

+++ Assets/Scripts/UpgradeMenu/Upgrade/UpgradeData.cs
@@ -9,7 +9,7 @@ using UnityEngine;
 public class UpgradeData
 {
     public StatType statType;
-    public int skillIncreaseAmount;
+    public float skillIncreaseAmount;
     public bool isPercentage;
 
 }
